Refactored the create and the get task by id operation to use the persistence layer
